requirements.txt
python-dotenv
requests
discord-py-interactions==4.4.0